+++ packages/email-wagmi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @web3modal/coinbase-wagmi-react-native
 
+## 1.4.2
+
+### Patch Changes
+
+- 1.4.2
+- Updated dependencies
+  - @web3modal/email-react-native@1.4.2
+
 ## 1.4.1
 
 ### Patch Changes